Reworked structure
